2014-07-25  Rong Xu  <xur@google.com>
	* libgcc/dyn-ipa.c (gcov_compute_random_module_groups): Fix for
        mingw build.
	(read_modu_groups_from_imports_files): Ditto.
	(__gcov_compute_module_groups): Ditto.
	* libgcc/libgcov-util.c (void gcov_set_verbose): Ditto.
	(set_use_modu_list): Ditto.
	(read_file_handler): Ditto.
	(ftw_read_file): Ditto.
	(myftw): Ditto.
	(gcov_read_profile_dir): Ditto.
	(gcov_profile_normalize): Ditto.
	* gcc/gcov-tool.c (unlink_gcda_file): Ditto.
	(unlink_profile_dir): Ditto.
	(gcov_output_files): Ditto.
	(module_name_hash_lookup): Ditto.
	(is_lipo_source_type): Ditto.
	(lipo_process_modu_list): Ditto.
	* gcc/l-ipo.c (type_addr_hash): Ditto.
	(equivalent_struct_types_for_tbaa): Ditto.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/google@213062 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc-4_9/gcc/l-ipo.c
@@ -530,7 +530,7 @@ type_addr_hash (const void *ent)
 {
   const struct type_ent *const entry
       = (const struct type_ent *) ent;
-  return (hashval_t) (long) entry->type;
+  return (hashval_t) (uintptr_t) entry->type;
 }
 
 /* Address equality function for type_ent.  */
@@ -858,14 +858,14 @@ equivalent_struct_types_for_tbaa (const_tree t1, const_tree t2)
   t1 = get_norm_type (t1);
   t2 = get_norm_type (t2);
 
-  key.type = (tree) (long) t1;
+  key.type = (tree) (uintptr_t) t1;
   slot = (struct type_ent **)
       htab_find_slot (l_ipo_type_tab, &key, NO_INSERT);
   if (!slot || !*slot)
     return -1;
   tent1 = *slot;
 
-  key.type = (tree) (long) t2;
+  key.type = (tree) (uintptr_t) t2;
   slot = (struct type_ent **)
       htab_find_slot (l_ipo_type_tab, &key, NO_INSERT);
   if (!slot || !*slot)
